+++ CHANGES.md
@@ -1,27 +1,33 @@
 # Release Note
 ## v1.3.2
-#### Bug Fixes:
+#### Bug Fixes
 - Log path: fix the default log path overwrites the custom log path when setting "manager.logger = True".
 #### Refactor
 - Adjust code structure.
 - Rewrite README.md.
+___
 ## v1.3.1
 - Delete a deprecated class.
 - Replace root logger with a separate logger.
 - Rename property "hot_keys" to "hotKeyList".
 - Change documents and some code comments.
+___
 ## v1.3.0
 - Currently, users can customize the log path.
 - Optimize code.
+___
 ## v1.2.0
 - Add logger.
 - Optimize code.
 - Attempt to fix a potential bug.
+___
 ## v1.1.1
 - Remove log message.
+___
 ## v1.1.0
 - Currently, the trigger function supports arguments.
 - No longer need to call manager.start() manually.
 - Fix multiple type hot key bug.
+___
 ## v1.0
 - The first version.